+++ fabric-menu-api-v1/src/testmod/java/net/fabricmc/fabric/test/menu/item/PositionedBagItem.java
@@ -32,7 +32,7 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.context.UseOnContext;
 import net.minecraft.world.level.Level;
 
-import net.fabricmc.fabric.test.menu.screen.PositionedBagMenu;
+import net.fabricmc.fabric.test.menu.menu.PositionedBagMenu;
 
 public class PositionedBagItem extends BagItem {
 	public PositionedBagItem(Properties properties) {
