+++ fabric-menu-api-v1/src/testmod/java/net/fabricmc/fabric/test/menu/menu/BoxMenu.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.test.menu.screen;
+package net.fabricmc.fabric.test.menu.menu;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.Container;
